+++ spec/prefixed.http.spec.js
@@ -29,7 +29,7 @@ describe( 'with http module', function() {
 		authProvider = require( './auth/mock.js' )( config );
 		passport = require( '../src/http/passport.js' )( config, authProvider, metrics );
 		middleware = require( '../src/http/middleware.js' )( config, metrics );
-		http = require( '../src/http/http.js' )( config, requestor, passport, middleware, metrics );
+		http = require( '../src/http/http.js' )( requestor, middleware, metrics );
 
 		authProvider.users = {};
 		passport.resetUserCheck();
@@ -51,7 +51,7 @@ describe( 'with http module', function() {
 			request = req;
 			res.status( statusCode ).send( response );
 		} );
-		http.start();
+		http.start( config, passport );
 	} );
 
 	describe( 'when getting a static file', function() {