+++ spec/socketio/skipped-auth.spec.js
@@ -21,10 +21,10 @@ describe( 'with socketio and no users', function() {
 		authProvider = require( '../auth/mock.js' )( config );
 		passport = require( '../../src/http/passport.js' )( config, authProvider, metrics );
 		middleware = require( '../../src/http/middleware.js' )( config, metrics );
-		http = require( '../../src/http/http.js' )( config, requestor, passport, middleware, metrics );
+		http = require( '../../src/http/http.js' )( requestor, middleware, metrics );
 		socket = require( '../../src/websocket/socket.js' )( config, http );
 
-		http.start();
+		http.start( config, passport );
 		socket.start( passport );
 		var onConnect = function() {
 			onConnect = function() {};