+++ spec/websocket/authenticated.spec.js
@@ -20,12 +20,12 @@ describe( 'with websocket and valid credentials', function() {
 		authProvider = require( '../auth/mock.js' )( config );
 		passport = require( '../../src/http/passport.js' )( config, authProvider, metrics );
 		middleware = require( '../../src/http/middleware.js' )( config, metrics );
-		http = require( '../../src/http/http.js' )( config, requestor, passport, middleware, metrics );
+		http = require( '../../src/http/http.js' )( requestor, middleware, metrics );
 		socket = require( '../../src/websocket/socket.js' )( config, http, middleware );
 
 		authProvider.users.admin = { name: 'admin', password: 'admin' };
 
-		http.start();
+		http.start( config, passport );
 		socket.start( passport );
 		
 		client = new WebSocketClient();