+++ spec/websocket/invalid-auth.spec.js
@@ -8,13 +8,6 @@ var config = {
 		socketio: true,
 		websocket: true
 	};
-// var authProvider = require( '../auth/mock.js' )( config );
-// var passport = require( '../../src/http/passport.js' )( config, authProvider, metrics );
-// var middleware = require( '../../src/http/middleware.js' )( config, metrics );
-// var http = require( '../../src/http/http.js' )( config, requestor, passport, middleware, metrics );
-// var socket = require( '../../src/websocket/socket.js' )( config, http, middleware );
-
-// authProvider.users[ 'test' ] = { user: 'torpald' };
 
 var authProvider, passport, middleware, http, socket;
 
@@ -26,12 +19,12 @@ describe( 'with websocket and bad credentials', function() {
 		authProvider = require( '../auth/mock.js' )( config );
 		passport = require( '../../src/http/passport.js' )( config, authProvider, metrics );
 		middleware = require( '../../src/http/middleware.js' )( config, metrics );
-		http = require( '../../src/http/http.js' )( config, requestor, passport, middleware, metrics );
+		http = require( '../../src/http/http.js' )( requestor, middleware, metrics );
 		socket = require( '../../src/websocket/socket.js' )( config, http, middleware );
 
 		authProvider.users[ 'test' ] = { user: 'torpald' };
 
-		http.start();
+		http.start( config, passport );
 		socket.start( passport );
 		client = new WebSocketClient();
 		client.connect(