+++ spec/websocket/noauth.spec.js
@@ -18,10 +18,10 @@ describe( 'with websocket and no auth strategy', function() {
 
 	before( function( done ) {
 		middleware = require( '../../src/http/middleware.js' )( config, metrics );
-		http = require( '../../src/http/http.js' )( config, requestor, undefined, middleware, metrics );
+		http = require( '../../src/http/http.js' )( requestor, middleware, metrics );
 		socket = require( '../../src/websocket/socket.js' )( config, http );
 
-		http.start();
+		http.start( config );
 		socket.start();
 		
 		client = new WebSocketClient();