+++ spec/ws.adapter.spec.js
@@ -35,7 +35,7 @@ describe( 'with socket adapter', function() {
 		authProvider = require( './auth/mock.js' )( config );
 		passport = require( '../src/http/passport.js' )( config, authProvider, metrics );
 		middleware = require( '../src/http/middleware.js' )( config, metrics );
-		http = require( '../src/http/http.js' )( config, requestor, passport, middleware, metrics );
+		http = require( '../src/http/http.js' )( requestor, middleware, metrics );
 		socket = require( '../src/websocket/socket.js' )( config, http, middleware );
 		socketAdapter = require( '../src/websocket/adapter.js' )( config, authProvider, socket, metrics );
 
@@ -67,7 +67,7 @@ describe( 'with socket adapter', function() {
 				env.reply( { data: { youSed: env.data.msg } } );
 			}
 		}, { topics: {} } );
-		http.start();
+		http.start( config, passport );
 		socket.start( passport );
 		ioClient = io( 'http://localhost:88988' );
 		ioClient.once( 'reconnect', check );