+++ src/http/httpEnvelope.js
@@ -49,6 +49,16 @@ HttpEnvelope.prototype.redirect = function( statusCode, url ) {
 
 HttpEnvelope.prototype.reply = function( envelope ) {
 	var code = envelope.statusCode || 200;
+	if( envelope.headers ) {
+		_.each( envelope.headers, function( v, k ) {
+			this._original.res.set( k, v );
+		}.bind( this ) );
+	}
+	if( envelope.cookies ) {
+		_.each( envelope.cookies, function( v, k ) {
+			this._original.res.cookie( k, v.value, v.options );
+		}.bind( this ) );
+	}
 	this._original.res.status( code ).send( envelope.data );
 };